Move storages classes into their own module
