+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/utils/storages.py
@@ -1,0 +1,25 @@
+{% if cookiecutter.cloud_provider == 'AWS' -%}
+from storages.backends.s3boto3 import S3Boto3Storage  # noqa E402
+
+
+class StaticRootS3Boto3Storage(S3Boto3Storage):
+    location = "static"
+    default_acl = "public-read"
+
+
+class MediaRootS3Boto3Storage(S3Boto3Storage):
+    location = "media"
+    file_overwrite = False
+{%- elif cookiecutter.cloud_provider == 'GCP' -%}
+from storages.backends.gcloud import GoogleCloudStorage  # noqa E402
+
+
+class StaticRootGoogleCloudStorage(GoogleCloudStorage):
+    location = "static"
+    default_acl = "publicRead"
+
+
+class MediaRootGoogleCloudStorage(GoogleCloudStorage):
+    location = "media"
+    file_overwrite = False
+{%- endif %}
